feat: Add checks to see if user exists

+++ src/controllers/user.ts
@@ -10,8 +10,15 @@ export const UserController = {
       select: userProjection(),
     });
 
+    if (!users) {
+      logger.error(
+        `method: ${req.method}, route: ${req.url}, message: Users not found`
+      );
+      res.status(404).json({ message: "Users not found" });
+    }
+
     logger.info(
-      `method: ${req.method}, route: ${req.url}, message: Get users succesful`
+      `method: ${req.method}, route: ${req.url}, message: Got users succesfully`
     );
     res.status(200).json({ data: users });
   }),
@@ -24,22 +31,44 @@ export const UserController = {
       select: userProjection(),
     });
 
+    if (!user) {
+      logger.error(
+        `method: ${req.method}, route: ${req.url}, message: User not found`
+      );
+      return res.status(404).json({ message: "User not found" });
+    }
+
     logger.info(
-      `method: ${req.method}, route: ${req.url}, message: Get user succesful -> ${user.email}`
+      `method: ${req.method}, route: ${req.url}, message: Got user succesfully -> ${user.email}`
     );
     res.status(200).json({ data: user });
   }),
 
   deleteUser: asyncHandler(async (req: Request, res: Response) => {
+    const userId = req.params.id;
+
+    const user = await prisma.user.findUnique({
+      where: {
+        id: userId,
+      },
+    });
+
+    if (!user) {
+      logger.error(
+        `method: ${req.method}, route: ${req.url}, message: User not found`
+      );
+      return res.status(404).json({ message: "User not found" });
+    }
+
     const deletedUser = await prisma.user.delete({
       where: {
-        id: req.params.id,
+        id: userId,
       },
       select: userProjection(),
     });
 
     logger.info(
-      `method: ${req.method}, route: ${req.url}, message: Find user succesful -> ${deletedUser.email}`
+      `method: ${req.method}, route: ${req.url}, message: User deleted succesfully -> ${deletedUser.email}`
     );
     res.status(200).json({ data: deletedUser });
   }),
